fix: back up the hotswap slot
The game checks if the slot being switched to is already selected and
does nothing if so. (For what it's worth, I did try changing the code
in memory to ignore these checks, but it didn't seem to work.) In
order to get the game to change slot, we write a different slot number
to the "current" slot in memory and then ask the game to change slot.

This doesn't swap the slot on screen and still works, which is
ideal. However, the game seems to overwrite the slot that is marked as
current when we do this. To remedy this issue, back up the slot before
telling the game to swap, then restore the slot afterwards.