feat: improve interactions with active HUD layout
Trying to preview different saved was more difficult than it
needed to be since after importing a layout, the user needed to
swap to another layout and then back to the original one in order
to see the changes. I changed this so that if a user imports a
layout to the currently active HUD slot, the plugin automatically
swaps to another HUD slot and back.

As a part of this also refined the general slot UI so that the
active slot is indicated in the slot selection UI for saving a
layout to game data.