feat: auto-restore pet hotbar after HUD changes
This is a much more stable and consistent fix for the pet hotbar problem (see #1). It now is performed via hooking the "set HUD" function. Whenever a change is made, the pet hotbar status is refreshed. This also has the side benefit of allowing for HUD manipulation while a pet hotbar is active.